requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
python-dotenv
openai
chromadb
tiktoken
faiss-cpu
unstructured
python-magic